pyproject.toml
[project]
name = "app"
version = "1.0.0"
description = "Svatky"
dependencies = [
    "flask",
    "pytest",
    "coverage",
    "flasgger",
    "python-dotenv",
    "Flask-Limiter",
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["app"]
